Better support for shouldComponentUpdate and re-rendering
I now better understand why shouldComponentUpdate works or does not work,
and added a large comment in Flux.hs with the details.  As part of that work,
the biggest change is that tuples of size two and three are specially coded
to check each component to see if it is the same javascript object as before.